Fix http://bugs.mysql.com/62534 by making the my.cnf innodb_max_dirty_pages_pct variable a double
Summary:
This makes innodb_max_dirty_pages_pct a double with min,default,max values
0.001, 75, 99.999. This also adds support for MYSQL_SYSVAR_DOUBLE.
"
Added more to the BUFFER POOL AND MEMORY section of SHOW INNODB STATUS
Percent pages dirty: 8.939         --> this is all n_dirty_pages / used_pages
+Percent all pages dirty: 0.195    --> this is n_dirty_pages / all-pages
+Max dirty pages percent: 75.000   --> this is innodb_max_dirty_pages_pct
"
Also made all of the buf use 3 digits of precision instead of 2 (%.2f -> %.3f)

Test Plan: mtr

Reviewers: nizamordulu

Reviewed By: nizamordulu

+++ storage/innodb_plugin/include/srv0srv.h
@@ -220,7 +220,7 @@ extern my_bool	srv_use_fast_checksums;
 extern ibool	srv_set_thread_priorities;
 extern int	srv_query_thread_priority;
 
-extern ulong	srv_max_buf_pool_modified_pct;
+extern double	srv_max_buf_pool_modified_pct;
 extern ulong	srv_max_purge_lag;
 
 extern ulong	srv_replication_delay;
